Fix fullscreen/sustained-activity capture + harden buffer math
Reported: fullscreen captures weren't working. Root cause: a surface that changes every frame (fullscreen video/game) never quiesces, so after the Initial frame the window never 'settled' and no images were ever saved.

- Add max_active_ms keyframe (default 5000ms): sustained activity now yields periodic captures instead of none.
- Implement fps_cap (was defined but unused): drop frames faster than the cap before the downsample pass; add Engine::frames_dropped().
- Refresh Windows window geometry (rect/client_rect/dpi/foreground) during capture so metadata stays correct across resize/fullscreen, instead of being frozen at start.
- Detect all-black frames and warn once (exclusive-fullscreen / DRM is an OS limitation WGC can't capture).
- Harden encode()/WorkingFrame buffer-size math to usize (avoid u32 overflow on multi-4K frames); GUI no longer panics on a poisoned mutex.
- Tests: sustained-change keyframe behavior (and the zero-capture repro), fps_cap rate limiting. All clippy configs + tests green on stable.

+++ src/detect/diff.rs
@@ -42,10 +42,11 @@ impl WorkingFrame {
                 let step_y = ((y1 - y0) / 16).max(1);
                 let mut py = y0;
                 while py < y1 {
-                    let row_off = (py * frame.stride) as usize;
+                    // usize math avoids u32 overflow on very large frames.
+                    let row_off = py as usize * frame.stride as usize;
                     let mut px = x0;
                     while px < x1 {
-                        let off = row_off + (px * 4) as usize;
+                        let off = row_off + px as usize * 4;
                         let b = frame.buffer.get(off).copied().unwrap_or(0) as u64;
                         let g = frame.buffer.get(off + 1).copied().unwrap_or(0) as u64;
                         let r = frame.buffer.get(off + 2).copied().unwrap_or(0) as u64;
